requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
async-timeout==5.0.1
attrs
babel
beautifulsoup4
bitsandbytes==0.45.3
bleach
Brotli
cached-property
certifi
cffi
charset-normalizer
comm
contourpy==1.3.1
cycler==0.12.1
datasets==3.3.2
debugpy
decorator==5.2.0
defusedxml
dill==0.3.8
et_xmlfile==2.0.0
exceptiongroup
executing
fastjsonschema
filelock==3.17.0
fire==0.7.0
fonttools==4.57.0
fqdn
frozenlist==1.5.0
fsspec==2024.12.0
h11
h2
hpack
httpcore
httpx
huggingface-hub==0.29.1
hyperframe
idna
importlib_metadata
importlib_resources
ipykernel
ipython
ipywidgets==8.1.5
isoduration
jedi
Jinja2
joblib==1.4.2
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets==3.0.13
jupyterlmod==5.2.1
kagglehub==0.3.9
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe
matplotlib==3.10.1
matplotlib-inline
mdurl==0.1.2
mistune
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nbclient
nbconvert
nbformat
nest_asyncio
networkx==3.4.2
notebook
notebook_shim
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.7
openpyxl==3.1.5
optimum==1.24.0
overrides
packaging
pandas==2.2.3
pandocfilters
parso
peft==0.14.0
pexpect
pickleshare
pillow==11.1.0
pkgutil_resolve_name
platformdirs
prometheus_client
prompt_toolkit
propcache==0.3.0
protobuf==5.29.3
psutil
ptyprocess
pure_eval
pyarrow==19.0.1
pycparser
pydantic==2.10.6
pydantic_core==2.27.2
Pygments
pyparsing==3.2.3
PySocks
python-dateutil
python-json-logger
pytz
PyYAML==6.0.2
pyzmq
referencing
regex==2024.11.6
requests
rfc3339_validator
rfc3986-validator
rich==13.9.4
rpds-py
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
Send2Trash
sentence-transformers==4.0.1
sentencepiece==0.2.0
six
sniffio
soupsieve
stack_data
sympy==1.13.1
termcolor==3.0.1
terminado
threadpoolctl==3.6.0
tinycss2
tokenizers==0.21.0
tomli
torch==2.6.0
tornado
tqdm==4.67.1
traitlets
transformers==4.49.0
triton==3.2.0
trl==0.16.0
types-python-dateutil
typing_extensions
typing_utils
tzdata==2025.1
uri-template
urllib3==2.3.0
wcwidth
webcolors
webencodings
websocket-client
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.3
zipp
